garbage_clean.domain: add restriction for startDeleteIfCan use case

+++ garbage_clean/domain/src/main/java/yin_kio/garbage_clean/domain/GarbageCleanerUseCases.kt
@@ -18,8 +18,10 @@ class GarbageCleanerUseCases(
     fun switchSelection(garbageType: GarbageType){
         deleteForm.switchSelection(garbageType)
     }
-    fun startDelete(){
-        files.delete(interpreter.interpret(deleteRequest))
+    fun startDeleteIfCan(){
+        if (deleteRequest.isNotEmpty()) {
+            files.delete(interpreter.interpret(deleteRequest))
+        }
     }
 
 }
